+++ CHANGELOG.md
@@ -1,5 +1,15 @@
+## Version 1.0.0.3
+* The following values now update in realtime:
+* Humidity
+* Pressure
+* Luminance
+* Current Power usage
+* Power used
+* Gas used
+
 ## Version 1.0.0.2
 * Changed clicking to touch and hold to cycle through tile values
+* Added Danish laguage file (da)
 
 ## Version 1.0.0.1
 * When a capability value causes an alarm the value will be shown on the tile